refactor(viewer): decompose the module into a package (no behavior change)
First step of the in-PR construction plan from review: split the
1,500-line viewer.py + 900-line inline template into a package with one
responsibility per module, exactly as suggested:

  trajectories/viewer/
    payload.py     ACP normalization, diagnostics, sidecar loading
    render.py      payload + packaged assets -> one self-contained page
    legacy.py      stream-json / Codex / raw-ACP inline renderers (verbatim)
    catalog.py     run discovery, cap, summaries
    sources.py     hf:// spec parsing and resolution
    server.py      single-run serving (confirm contract) + browse mode
    assets/        theme.css / viewer.css / viewer.js / template.html
    __main__.py    python -m entrypoint

The split is mechanical (functions moved verbatim by line range); the only
rewritten body is _load_template(), which now assembles the page from the
split assets. benchflow.trajectories.viewer keeps its full import surface
via __init__ re-exports — all 394 viewer-battery tests pass unmodified,
and the wheel ships the assets subtree.

Next steps per the plan: typed payload contract (models.py), the real
SkillsBench theme, then the catalog information architecture.
