+++ src/benchflow/trajectories/viewer/assets/template.html
@@ -1,0 +1,35 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+<meta charset="utf-8">
+<meta name="viewport" content="width=device-width, initial-scale=1">
+<title>__BENCHFLOW_TITLE__</title>
+<link rel="icon" href="data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 16 16'%3E%3Ccircle cx='8' cy='8' r='6' fill='%230a0a0a'/%3E%3C/svg%3E">
+<style>
+/*__BENCHFLOW_THEME_CSS__*/
+/*__BENCHFLOW_VIEWER_CSS__*/
+</style>
+</head>
+<body>
+<div id="app">
+  <aside id="side" class="hidden" aria-label="runs"></aside>
+  <div id="content">
+    <div class="wordmark">
+      <svg viewBox="0 0 514 512" fill="currentColor" aria-hidden="true"><path fill-rule="evenodd" clip-rule="evenodd" d="M445.422 66.4597L511.882 0L389.022 293.965L295.129 387.859L0 511.882L69.3042 442.577L81.0101 454.283L89.0554 446.238L77.3493 434.532L130.65 381.232L162.469 413.051L170.514 405.006L138.695 373.187L191.995 319.887L203.701 331.593L211.746 323.547L200.04 311.841L253.34 258.541L285.16 290.36L293.205 282.315L261.386 250.496L314.686 197.196L326.392 208.902L334.437 200.856L322.731 189.15L376.031 135.85L407.851 167.669L415.896 159.624L384.077 127.805L437.377 74.5049L449.083 86.2108L457.128 78.1656L445.422 66.4597ZM399.127 389.865V299.369L513.197 26.4333V503.935L399.127 389.865ZM391.061 397.931L505.132 512.001H29.1594L300.605 397.931H391.061Z"/></svg>
+      <span class="brand">BenchFlow</span>
+      <span class="app">trajectory viewer</span>
+    </div>
+    <header id="hdr"></header>
+    <nav id="tabs" role="tablist"></nav>
+    <div id="toolbar"></div>
+    <main id="view-trace" role="tabpanel"></main>
+    <main id="view-verifier" role="tabpanel" class="hidden"></main>
+    <main id="view-metrics" role="tabpanel" class="hidden"></main>
+  </div>
+</div>
+<script id="bf-payload" type="application/json">__BENCHFLOW_PAYLOAD__</script>
+<script>
+//__BENCHFLOW_VIEWER_JS__
+</script>
+</body>
+</html>
