+++ src/benchflow/trajectories/viewer/legacy.py
@@ -1,21 +1,15 @@
-"""Trajectory viewer — renders Claude Code stream-json, Codex sessions, and ACP JSONL as HTML.
+"""Inline-CSS renderers: Claude Code stream-json, Codex sessions, raw ACP files.
 
-Works with trial directories (`turn*.txt` or `trajectory/acp_trajectory.jsonl`),
-with a raw session JSONL file, and — for a directory *of* rollouts — as a
-browsable multi-run server (sidebar + JSON API, see ``_serve_browse``). ACP
-rollout directories render through the interactive ``viewer_template.html``
-page; stream-json/Codex/raw-file paths keep the shared inline-CSS renderers.
-No ATIF conversion.
+The pre-package renderers, moved verbatim. Their server-rendered output is
+pinned by the jsonl-session and confirm-flow tests.
 """
 
 import html
 import json
-import os
-import sys
-from datetime import datetime
-from importlib import resources
 from pathlib import Path
-from typing import Any
+
+from .payload import _load_result_json, _parse_jsonl
+from .render import _render_acp_trajectory
 
 _THINKING_PREVIEW = 600  # max chars for thinking block preview
 _ARGS_PREVIEW = 300  # max chars for tool args display
@@ -364,20 +358,6 @@ def _user_prompt_html(text: str) -> str:
     )
 
 
-def _parse_jsonl(text: str) -> list[dict]:
-    events = []
-    for line in text.splitlines():
-        if not line.strip():
-            continue
-        try:
-            parsed = json.loads(line)
-        except json.JSONDecodeError:
-            continue
-        if isinstance(parsed, dict):
-            events.append(parsed)
-    return events
-
-
 def render_rollout(rollout_dir: Path, prompts: list[str] | None = None) -> str:
     """Render a full trial (multiple turns) as HTML.
 
@@ -481,445 +461,6 @@ def render_rollout(rollout_dir: Path, prompts: list[str] | None = None) -> str:
 # the self-contained vanilla-JS page renders it client-side. Raw session
 # FILES keep the inline `_render_acp_events` renderer below (its output is
 # pinned server-side by the jsonl-session tests).
-
-_TEMPLATE_NAME = "viewer_template.html"
-_PAYLOAD_PLACEHOLDER = "__BENCHFLOW_PAYLOAD__"
-_TITLE_PLACEHOLDER = "__BENCHFLOW_TITLE__"
-
-# result.json diagnostic blocks surfaced in the header banner. The live list
-# derives from DIAGNOSTIC_REGISTRY so new diagnostics (e.g. behavior flags)
-# appear without this module drifting; the fallback pins the 0.7.4 set.
-_DIAGNOSTIC_KEYS_FALLBACK = (
-    "idle_timeout_info",
-    "agent_timeout_info",
-    "sandbox_startup_info",
-    "transport_error_info",
-    "verifier_timeout_info",
-    "api_error_info",
-    "suspected_api_error_info",
-)
-
-
-def _diagnostic_keys() -> tuple[str, ...]:
-    """Diagnostic block keys, in registry display order."""
-    try:
-        from benchflow.diagnostics import DIAGNOSTIC_REGISTRY
-    except Exception:  # pragma: no cover — registry is core; stay lenient
-        return _DIAGNOSTIC_KEYS_FALLBACK
-    return tuple(d.field for d in DIAGNOSTIC_REGISTRY)
-
-
-def _load_json(path: Path) -> Any:
-    # errors="replace": binary/mis-encoded sidecars degrade to a JSON parse
-    # failure (→ None) instead of an unhandled UnicodeDecodeError.
-    try:
-        return json.loads(path.read_text(encoding="utf-8", errors="replace"))
-    except (json.JSONDecodeError, OSError):
-        return None
-
-
-def _read_text(path: Path) -> str | None:
-    try:
-        return path.read_text(encoding="utf-8", errors="replace")
-    except OSError:
-        return None
-
-
-def _tool_content_texts(content: Any) -> list[str]:
-    """Flatten ACP tool_call content blocks to plain strings.
-
-    Three shapes occur in captures: the canonical nested block
-    ``{"type": "content", "content": {"type": "text", "text": ...}}``, a flat
-    ``{"text": ...}``, and edit-tool diff blocks
-    ``{"type": "diff", "path": ..., "oldText": ..., "newText": ...}``.
-    """
-    texts: list[str] = []
-    if not isinstance(content, list):
-        return texts
-    for item in content:
-        if not isinstance(item, dict):
-            texts.append(str(item))
-            continue
-        inner = item.get("content")
-        if isinstance(inner, dict) and "text" in inner:
-            texts.append(str(inner.get("text", "")))
-        elif "text" in item:
-            texts.append(str(item.get("text", "")))
-        elif item.get("type") == "diff":
-            texts.append(
-                f"diff {item.get('path', '')}\n"
-                f"--- old\n{item.get('oldText') or ''}\n"
-                f"+++ new\n{item.get('newText') or ''}"
-            )
-        else:
-            texts.append(json.dumps(item, ensure_ascii=False))
-    return [t for t in texts if t]
-
-
-def _parse_ts(value: Any) -> float | None:
-    """Lenient timestamp → epoch seconds; None for absent/unparseable.
-
-    Accepts numeric epochs and anything ``datetime.fromisoformat`` reads
-    (ISO-8601, and the space-separated ``str(datetime.now())`` form used by
-    result.json's started_at).
-    """
-    if isinstance(value, bool):
-        return None
-    if isinstance(value, (int, float)):
-        return float(value)
-    if isinstance(value, str):
-        try:
-            return datetime.fromisoformat(value).timestamp()
-        except ValueError:
-            return None
-    return None
-
-
-def _normalize_steps(events: list[dict], prompts: list[str] | None) -> list[dict]:
-    """Project raw ACP events onto renderer steps with server-side numbering.
-
-    Prompt labels ("PROMPT 1") are computed here, not client-side, and the
-    ``prompts`` list is embedded only when the trajectory has no inline
-    user_message events — embedding both would duplicate prompt text in the
-    emitted page (pinned by TestViewerCompatibility).
-
-    Timestamps are forward-compatible passthrough: today's captures carry
-    none (verified across the HF ground-truth uploads), but when the
-    capture-side proposal lands (``ts`` on text events, ``started_at`` /
-    ``finished_at`` on tool calls — benchflow#1033) steps gain ``t`` (epoch)
-    and tools ``dur`` (seconds), and the renderer shows a timeline with no
-    further changes.
-    """
-    steps: list[dict] = []
-    prompt_counter = 0
-
-    def add(kind: str, event: dict | None = None, **fields: Any) -> None:
-        step = {"i": len(steps) + 1, "kind": kind, **fields}
-        if event is not None:
-            if kind == "tool":
-                started = _parse_ts(event.get("started_at") or event.get("ts"))
-                finished = _parse_ts(event.get("finished_at"))
-                if started is not None:
-                    step["t"] = started
-                    if finished is not None and finished >= started:
-                        step["dur"] = finished - started
-            else:
-                ts = _parse_ts(event.get("ts"))
-                if ts is not None:
-                    step["t"] = ts
-        steps.append(step)
-
-    has_inline_prompts = any(e.get("type") == "user_message" for e in events)
-    if not has_inline_prompts:
-        for text in prompts or []:
-            prompt_counter += 1
-            add("prompt", label=f"PROMPT {prompt_counter}", text=str(text))
-
-    for event in events:
-        etype = event.get("type", "")
-        if etype == "user_message":
-            prompt_counter += 1
-            add(
-                "prompt",
-                event,
-                label=f"PROMPT {prompt_counter}",
-                text=str(event.get("text", "")),
-            )
-        elif etype == "agent_message":
-            add("message", event, text=str(event.get("text", "")))
-        elif etype == "agent_thought":
-            add("thought", event, text=str(event.get("text", "")))
-        elif etype == "tool_call":
-            add(
-                "tool",
-                event,
-                tool={
-                    "id": event.get("tool_call_id", ""),
-                    "kind": event.get("kind", "other"),
-                    "title": event.get("title", ""),
-                    "status": event.get("status", ""),
-                    "content": _tool_content_texts(event.get("content")),
-                },
-            )
-        elif etype == "agent_timeout":
-            # Coerce to a list of strings — a crafted non-list value would
-            # otherwise reach the client and break the renderer.
-            pending_raw = event.get("pending_tool_call_ids")
-            add(
-                "timeout",
-                event,
-                timeout={
-                    "reason": event.get("reason", ""),
-                    "timeout_sec": event.get("timeout_sec"),
-                    "pending": [str(x) for x in pending_raw]
-                    if isinstance(pending_raw, list)
-                    else [],
-                    "complete": event.get("terminal_trajectory_complete"),
-                },
-            )
-        else:
-            # Unknown/future event types render as a generic card, never crash.
-            add(
-                "unknown",
-                event,
-                type=str(etype),
-                text=json.dumps(event, ensure_ascii=False)[:2000],
-            )
-    return steps
-
-
-def _build_meta(
-    result_data: dict, timing: dict | None, steps: list[dict]
-) -> dict[str, Any]:
-    rewards = result_data.get("rewards")
-    reward = rewards.get("reward") if isinstance(rewards, dict) else None
-    agent_result = result_data.get("agent_result")
-    usage = agent_result if isinstance(agent_result, dict) else {}
-
-    errors: list[dict[str, str]] = []
-    if result_data.get("error"):
-        errors.append(
-            {
-                "label": str(result_data.get("error_category") or "error"),
-                "text": str(result_data["error"]),
-            }
-        )
-    if result_data.get("verifier_error"):
-        errors.append(
-            {
-                "label": str(
-                    result_data.get("verifier_error_category") or "verifier error"
-                ),
-                "text": str(result_data["verifier_error"]),
-            }
-        )
-    if result_data.get("export_error"):
-        errors.append(
-            {"label": "export error", "text": str(result_data["export_error"])}
-        )
-    # Diagnostics that ride along an actual error render as error banners;
-    # on an otherwise-clean rollout they are behavior flags (e.g. #1025's
-    # chat-only completion) and render neutrally. All three error channels
-    # count as "actual error".
-    has_error = bool(
-        result_data.get("error")
-        or result_data.get("verifier_error")
-        or result_data.get("export_error")
-    )
-    for key in _diagnostic_keys():
-        value = result_data.get(key)
-        if value:
-            errors.append(
-                {
-                    "label": key.removesuffix("_info").replace("_", " "),
-                    "text": json.dumps(value, ensure_ascii=False, default=str)[:2000],
-                    "level": "error" if has_error else "info",
-                }
-            )
-
-    return {
-        "task_name": result_data.get("task_name"),
-        "agent_name": result_data.get("agent_name") or result_data.get("agent"),
-        "model": result_data.get("model"),
-        "skill_mode": result_data.get("skill_mode"),
-        "reward": reward,
-        "usage": usage,
-        "counts": {
-            "prompts": sum(1 for s in steps if s["kind"] == "prompt"),
-            "messages": sum(1 for s in steps if s["kind"] == "message"),
-            "thoughts": sum(1 for s in steps if s["kind"] == "thought"),
-            "tools": sum(1 for s in steps if s["kind"] == "tool"),
-        },
-        "timing": timing,
-        "duration_sec": (timing or {}).get("total"),
-        "errors": errors,
-        "trajectory_source": result_data.get("trajectory_source"),
-        "partial_trajectory": result_data.get("partial_trajectory"),
-        "started_at": result_data.get("started_at"),
-        "finished_at": result_data.get("finished_at"),
-    }
-
-
-def _load_verifier(rollout_dir: Path) -> dict[str, Any]:
-    vdir = rollout_dir / "verifier"
-    reward = _read_text(vdir / "reward.txt")
-    ctrf_tests = None
-    ctrf = _load_json(vdir / "ctrf.json")
-    if isinstance(ctrf, dict):
-        raw_tests = (ctrf.get("results") or {}).get("tests")
-        if isinstance(raw_tests, list):
-            ctrf_tests = [
-                {
-                    "name": str(t.get("name", "")),
-                    "status": str(t.get("status", "")),
-                    "duration": t.get("duration"),
-                }
-                for t in raw_tests
-                if isinstance(t, dict)
-            ]
-    return {
-        "reward": reward.strip() if reward else None,
-        "stdout": _read_text(vdir / "test-stdout.txt"),
-        "stderr": _read_text(vdir / "test-stderr.txt"),
-        "ctrf": ctrf_tests,
-    }
-
-
-def _load_template() -> str:
-    return (resources.files("benchflow.trajectories") / _TEMPLATE_NAME).read_text(
-        encoding="utf-8"
-    )
-
-
-def _safe_json(obj: Any) -> str:
-    """Serialize untrusted payload data to UTF-8-safe JSON text.
-
-    Lone surrogates (possible in crafted trajectory text) survive
-    ``json.dumps(ensure_ascii=False)`` but crash utf-8 encoding later in
-    ``serve()``; replace them here.
-    """
-    data = json.dumps(obj, ensure_ascii=False, default=str)
-    return data.encode("utf-8", errors="replace").decode("utf-8")
-
-
-def _render_shell(title: str, boot: dict[str, Any]) -> str:
-    """Inject a boot document into the template (script-breakout escaped)."""
-    page = _load_template()
-    page = page.replace(_TITLE_PLACEHOLDER, html.escape(title), 1)
-    return page.replace(_PAYLOAD_PLACEHOLDER, _safe_json(boot).replace("</", "<\\/"), 1)
-
-
-def _build_acp_payload(rollout_dir: Path, prompts: list[str] | None) -> dict[str, Any]:
-    """Assemble the renderer payload for one canonical ACP rollout dir."""
-    acp_path = rollout_dir / "trajectory" / "acp_trajectory.jsonl"
-    try:
-        text = acp_path.read_text(encoding="utf-8", errors="replace")
-    except OSError:
-        text = ""
-    events = _parse_jsonl(text)
-
-    if prompts is None:
-        loaded = _load_json(rollout_dir / "prompts.json")
-        prompts = loaded if isinstance(loaded, list) else None
-
-    result_data = _load_result_json(rollout_dir)
-    timing = _load_json(rollout_dir / "timing.json")
-    if not isinstance(timing, dict):
-        embedded = result_data.get("timing")
-        timing = embedded if isinstance(embedded, dict) else None
-
-    steps = _normalize_steps(events, prompts)
-    return {
-        "schema_version": 1,
-        "rollout_name": rollout_dir.name,
-        "meta": _build_meta(result_data, timing, steps),
-        "steps": steps,
-        "verifier": _load_verifier(rollout_dir),
-    }
-
-
-def _is_acp_rollout_dir(path: Path) -> bool:
-    return (path / "trajectory" / "acp_trajectory.jsonl").exists()
-
-
-def _runs_cap() -> int:
-    """Browse-mode run cap (``BENCHFLOW_VIEWER_MAX_RUNS`` overrides, default 500)."""
-    try:
-        return max(1, int(os.environ.get("BENCHFLOW_VIEWER_MAX_RUNS", "500")))
-    except ValueError:
-        return 500
-
-
-def _discover_rollouts(
-    base: Path, max_depth: int = 4, cap: int | None = None
-) -> list[str]:
-    """Relative paths of ACP rollout dirs under ``base``, sorted, capped.
-
-    The returned ids double as the ``/api/rollout?id=`` whitelist: an id is
-    only ever resolved by exact membership here, so crafted ids (``../``
-    traversal and the like) can never reach the filesystem. Directory
-    symlinks under ``base`` are followed — serving a directory implies
-    trusting what it links to.
-    """
-    if cap is None:
-        cap = _runs_cap()
-    found: list[str] = []
-
-    def walk(d: Path, depth: int) -> None:
-        if len(found) >= cap:
-            return
-        if _is_acp_rollout_dir(d):
-            found.append(d.relative_to(base).as_posix())
-            return  # rollout dirs don't nest
-        if depth >= max_depth:
-            return
-        try:
-            children = sorted(p for p in d.iterdir() if p.is_dir())
-        except OSError:
-            return
-        for child in children:
-            if child.name.startswith("."):
-                continue
-            walk(child, depth + 1)
-
-    try:
-        top = sorted(p for p in base.iterdir() if p.is_dir())
-    except OSError:
-        return found
-    for child in top:
-        if not child.name.startswith("."):
-            walk(child, 1)
-    return found
-
-
-def _resolve_browse_rollout(base: Path, rid: str | None) -> Path | None:
-    """Resolve an ``/api/rollout`` id strictly by whitelist membership.
-
-    The id is never interpreted as a path unless a fresh scan discovered it,
-    so crafted ids (``../`` traversal, absolute paths) return ``None`` even
-    when the traversed-to path exists and is a real rollout.
-    """
-    if rid is None or rid not in set(_discover_rollouts(base)):
-        return None
-    return base / rid
-
-
-def _rollout_summary(base: Path, rel_id: str) -> dict[str, Any]:
-    """Sidebar row for one rollout: identity + score at a glance."""
-    d = base / rel_id
-    result_data = _load_result_json(d)
-    rewards = result_data.get("rewards")
-    reward = rewards.get("reward") if isinstance(rewards, dict) else None
-    return {
-        "id": rel_id,
-        "name": d.name,
-        "task_name": result_data.get("task_name") or d.name,
-        "agent_name": result_data.get("agent_name") or result_data.get("agent"),
-        "model": result_data.get("model"),
-        "reward": reward,
-        "has_error": bool(
-            result_data.get("error") or result_data.get("verifier_error")
-        ),
-        "skill_mode": result_data.get("skill_mode"),
-    }
-
-
-def _render_acp_trajectory(
-    rollout_dir: Path, acp_path: Path, prompts: list[str] | None
-) -> str:
-    """Render a canonical ACP rollout as a self-contained interactive page.
-
-    Trajectory content is untrusted input: it travels as JSON data (``</``
-    escaped so it cannot break out of the script tag) and the template
-    renders it exclusively via ``textContent``.
-    """
-    payload = _build_acp_payload(rollout_dir, prompts)
-    return _render_shell(rollout_dir.name, {"mode": "single", "payload": payload})
-
-
-def _load_result_json(rollout_dir: Path) -> dict:
-    parsed = _load_json(rollout_dir / "result.json")
-    return parsed if isinstance(parsed, dict) else {}
 
 
 def _render_acp_events(
@@ -1200,317 +741,3 @@ def _stream_json_page(
 
 # Sidecar files the viewer reads from a rollout dir; hf:// resolution fetches
 # only these (large llm_trajectory.jsonl / trainer exports are skipped).
-_HF_VIEWER_FILES = (
-    "result.json",
-    "timing.json",
-    "prompts.json",
-    "trajectory/acp_trajectory.jsonl",
-    "verifier/*",
-)
-
-
-def _parse_hf_spec(spec: str) -> tuple[str, str | None, str]:
-    """Split ``hf://org/name[@revision][/subpath]`` into its parts.
-
-    Tolerates ``hf:/`` (a Path round-trip collapses the double slash) and
-    returns ``(repo_id, revision, subpath)``.
-    """
-    raw = spec.split(":", 1)[1].lstrip("/")
-    parts = [p for p in raw.split("/") if p]
-    if len(parts) < 2:
-        raise ValueError(
-            f"invalid HF dataset spec {spec!r} — expected hf://<org>/<name>[/subpath]"
-        )
-    repo_id = "/".join(parts[:2])
-    revision = None
-    if "@" in parts[1]:
-        name, _, revision = parts[1].partition("@")
-        repo_id = f"{parts[0]}/{name}"
-    return repo_id, revision, "/".join(parts[2:])
-
-
-def _resolve_hf_source(spec: str) -> Path:
-    """Materialize the viewer-relevant slice of an HF dataset locally.
-
-    Downloads (into the shared huggingface_hub cache, so repeat views are
-    incremental) only the files the viewer renders — trajectories plus
-    result/timing/prompts/verifier sidecars — and returns the local
-    directory to serve.
-    """
-    try:
-        from huggingface_hub import snapshot_download
-    except ModuleNotFoundError:
-        print("huggingface_hub is required for hf:// sources")
-        sys.exit(1)
-
-    repo_id, revision, subpath = _parse_hf_spec(spec)
-    scope = f"{subpath.rstrip('/')}/" if subpath else ""
-    patterns = []
-    for name in _HF_VIEWER_FILES:
-        patterns.append(f"{scope}{name}")
-        patterns.append(f"{scope}**/{name}")
-    print(f"Fetching {repo_id}" + (f" @ {revision}" if revision else "") + " …")
-    local = snapshot_download(
-        repo_id=repo_id,
-        repo_type="dataset",
-        revision=revision,
-        allow_patterns=patterns,
-    )
-    root = Path(local) / subpath if subpath else Path(local)
-    if not root.is_dir():
-        print(f"No such path in dataset {repo_id}: {subpath}")
-        sys.exit(1)
-    return root
-
-
-def serve(
-    rollout_path: str,
-    port: int = 8888,
-    prompts: list[str] | None = None,
-    confirm: bool = False,
-    redaction_summary: str | None = None,
-) -> str | None:
-    """Serve a trial directory, a session JSONL file, a directory of runs,
-    or an ``hf://`` dataset source.
-
-    ``hf://<org>/<name>[/subpath]`` fetches the viewer-relevant slice of a
-    HuggingFace trajectory dataset (e.g. the community ground-truth uploads)
-    into the local HF cache and serves it like any local directory.
-
-    Single trajectories (a rollout directory, or a session JSONL file) keep
-    the one-page behavior, including the ``confirm=True`` Approve/Reject
-    contract documented on :func:`_serve_single`. A directory that is not
-    itself a rollout but contains rollout directories is served in browse
-    mode instead: a run sidebar plus ``/api/rollout?id=…`` endpoints the page
-    uses to load traces dynamically (see :func:`_serve_browse`).
-
-    ``confirm`` needs exactly one trajectory to approve, so combining it with
-    a multi-run directory is an error.
-    """
-    if isinstance(rollout_path, str) and rollout_path.startswith("hf:"):
-        rollout_path = str(_resolve_hf_source(rollout_path))
-    path = Path(rollout_path)
-    if (
-        path.is_dir()
-        and not _is_acp_rollout_dir(path)
-        and not any(path.glob("turn*.txt"))
-    ):
-        cap = _runs_cap()
-        rollouts = _discover_rollouts(path, cap=cap + 1)
-        if rollouts:
-            capped = len(rollouts) > cap
-            n_runs = min(len(rollouts), cap)
-            if confirm:
-                print(
-                    "--confirm needs a single rollout or session file, but "
-                    f"{path} is a directory of {n_runs}{'+' if capped else ''} runs"
-                )
-                sys.exit(1)
-            _serve_browse(path, port, n_runs=n_runs, capped=capped)
-            return None
-    return _serve_single(path, port, prompts, confirm, redaction_summary)
-
-
-def _serve_single(
-    path: Path,
-    port: int,
-    prompts: list[str] | None,
-    confirm: bool,
-    redaction_summary: str | None,
-) -> str | None:
-    """Serve one trial directory or session JSONL file as a web page.
-
-    With ``confirm=True`` the page carries an Approve/Reject bar posting to
-    ``/decision``; the first valid decision shuts the server down and is
-    returned as ``"approved"`` or ``"rejected"`` after printing a
-    machine-readable ``DECISION: <value>`` line to stdout. Without it the
-    server runs until Ctrl+C and the return value is ``None`` — exactly the
-    pre-confirm behavior (no bar, no POST endpoint).
-
-    ``redaction_summary`` is an optional caller-composed line (e.g. ``"2 API
-    keys, 1 bearer token"``) rendered inside the confirm bar so the reviewer
-    sees what upload-time redaction would mask. Presentation-only; it has no
-    effect without ``confirm=True``.
-    """
-    import threading
-    from http.server import HTTPServer, SimpleHTTPRequestHandler
-
-    write_sidecar = False
-    if path.is_file():
-        html_content = render_jsonl_file(path)
-    elif path.is_dir():
-        html_content = render_rollout(path, prompts)
-        write_sidecar = True
-    else:
-        print(f"Not a file or directory: {path}")
-        sys.exit(1)
-
-    if html_content == _NO_TRAJECTORIES_HTML:
-        # Don't write a blank trajectory.html into an unrelated directory or
-        # start a server for nothing — fail fast like the not-a-directory path.
-        print(f"No trajectories found in {path}")
-        sys.exit(1)
-    if write_sidecar:
-        # The sidecar stays the plain page: the confirm bar is a one-shot
-        # interaction against this live server, not part of the artifact.
-        (path / "trajectory.html").write_text(html_content)
-    if confirm:
-        html_content = _inject_confirm_bar(html_content, redaction_summary)
-
-    print(f"Trajectory viewer: http://localhost:{port}")
-    print(f"Trial: {path}")
-    if confirm:
-        print("Waiting for Approve / Not this one in the browser (Ctrl+C to stop)\n")
-    else:
-        print("Press Ctrl+C to stop\n")
-
-    decision: str | None = None
-
-    class Handler(SimpleHTTPRequestHandler):
-        def do_GET(self):
-            self.send_response(200)
-            self.send_header("Content-Type", "text/html; charset=utf-8")
-            self.end_headers()
-            self.wfile.write(html_content.encode())
-
-        def log_message(self, format, *args):
-            pass
-
-    handler_cls: type[SimpleHTTPRequestHandler] = Handler
-
-    if confirm:
-
-        class ConfirmHandler(Handler):
-            def do_POST(self):
-                nonlocal decision
-                if self.path != "/decision":
-                    self.send_error(404)
-                    return
-                length = int(self.headers.get("Content-Length") or 0)
-                body = self.rfile.read(length).decode("utf-8", "replace").strip()
-                if body not in ("approve", "reject"):
-                    self.send_error(400, "Body must be 'approve' or 'reject'")
-                    return
-                decision = "approved" if body == "approve" else "rejected"
-                self.send_response(200)
-                self.send_header("Content-Type", "text/plain; charset=utf-8")
-                self.end_headers()
-                self.wfile.write(decision.encode())
-                # shutdown() blocks until serve_forever returns, and this
-                # handler runs inside that loop — stop from a helper thread.
-                threading.Thread(target=server.shutdown, daemon=True).start()
-
-        handler_cls = ConfirmHandler
-
-    server = HTTPServer(("localhost", port), handler_cls)
-    try:
-        server.serve_forever()
-    except KeyboardInterrupt:
-        print("\nStopped.")
-        return None
-    finally:
-        server.server_close()
-
-    if decision is not None:
-        print(f"DECISION: {decision}", flush=True)
-    return decision
-
-
-def _serve_browse(base: Path, port: int, n_runs: int, capped: bool = False) -> None:
-    """Multi-rollout browser: sidebar shell + JSON API, rescanned per request."""
-    from http.server import HTTPServer, SimpleHTTPRequestHandler
-    from urllib.parse import parse_qs, urlsplit
-
-    class Handler(SimpleHTTPRequestHandler):
-        def _send(
-            self,
-            status: int,
-            content_type: str,
-            body: bytes,
-            headers: tuple[tuple[str, str], ...] = (),
-        ) -> None:
-            self.send_response(status)
-            self.send_header("Content-Type", content_type)
-            for name, value in headers:
-                self.send_header(name, value)
-            self.end_headers()
-            self.wfile.write(body)
-
-        def _scan(self) -> tuple[list[str], bool]:
-            """Fresh id list plus whether the cap truncated it."""
-            cap = _runs_cap()
-            ids = _discover_rollouts(base, cap=cap + 1)
-            return ids[:cap], len(ids) > cap
-
-        def do_GET(self):
-            parsed = urlsplit(self.path)
-            if parsed.path == "/":
-                ids, capped = self._scan()
-                shell = _render_shell(
-                    base.name,
-                    {
-                        "mode": "browse",
-                        "capped": capped,
-                        "rollouts": [_rollout_summary(base, r) for r in ids],
-                    },
-                )
-                self._send(
-                    200,
-                    "text/html; charset=utf-8",
-                    shell.encode("utf-8", errors="replace"),
-                )
-            elif parsed.path == "/api/rollouts":
-                ids, truncated = self._scan()
-                body = _safe_json([_rollout_summary(base, r) for r in ids])
-                self._send(
-                    200,
-                    "application/json; charset=utf-8",
-                    body.encode("utf-8", errors="replace"),
-                    # API consumers must be able to detect truncation too; the
-                    # body stays a bare list for backward compatibility.
-                    headers=(("X-BenchFlow-Capped", "1"),) if truncated else (),
-                )
-            elif parsed.path == "/api/rollout":
-                rid = (parse_qs(parsed.query).get("id") or [None])[0]
-                rollout_dir = _resolve_browse_rollout(base, rid)
-                if rollout_dir is None:
-                    self._send(
-                        404,
-                        "application/json; charset=utf-8",
-                        b'{"error": "unknown rollout id"}',
-                    )
-                    return
-                body = _safe_json(_build_acp_payload(rollout_dir, None))
-                self._send(
-                    200,
-                    "application/json; charset=utf-8",
-                    body.encode("utf-8", errors="replace"),
-                )
-            else:
-                self._send(404, "text/plain; charset=utf-8", b"not found")
-
-        def log_message(self, format, *args):
-            pass
-
-    runs_desc = f"{n_runs} runs"
-    if capped:
-        runs_desc = f"first {n_runs} runs (capped — raise BENCHFLOW_VIEWER_MAX_RUNS)"
-    print(f"Trajectory browser: http://localhost:{port}")
-    print(f"Scanning: {base} ({runs_desc})")
-    print("Press Ctrl+C to stop\n")
-
-    server = HTTPServer(("localhost", port), Handler)
-    try:
-        server.serve_forever()
-    except KeyboardInterrupt:
-        print("\nStopped.")
-    finally:
-        server.server_close()
-
-
-if __name__ == "__main__":
-    if len(sys.argv) < 2:
-        print("Usage: python -m benchflow.viewer <rollout_dir_or_jsonl> [port]")
-        sys.exit(1)
-    port = int(sys.argv[2]) if len(sys.argv) > 2 else 8888
-    serve(sys.argv[1], port)
